backend/requirements.txt
# Core dependencies
fastapi>=0.68.1
uvicorn==0.15.0
pydantic==1.10.13
pydantic-settings==0.2.5

# API Dependencies
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML Dependencies (Core)
scikit-learn>=0.24.2
numpy>=1.19.5
pandas>=1.3.4
joblib>=1.1.0

# ML Tools
optuna>=2.10.0
nltk>=3.6.5

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0 
pytest-cov==4.1.0
httpx==0.24.1
coverage==7.2.7

# Utilities
python-dotenv==0.19.2
requests==2.31.0
aiohttp==3.8.5
async-timeout==4.0.2